# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lumen API",
        "description": "API for the Lumen project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/planes-estudio/carrera/{id_carrera}": {
            "get": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Obtener el plan de estudio de una carrera",
                "operationId": "0aa8bfdb29b3add64e13b7ca4d66b59e",
                "parameters": [
                    {
                        "name": "id_carrera",
                        "in": "path",
                        "description": "ID de la carrera",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan de estudio obtenido con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_carrera": {
                                                "type": "integer"
                                            },
                                            "id_modalidad": {
                                                "type": "integer"
                                            },
                                            "materias": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id_materia": {
                                                            "type": "integer"
                                                        },
                                                        "id_cat_nivel": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No se encontró un plan de estudio para la carrera especificada"
                    },
                    "404": {
                        "description": "La carrera no existe"
                    }
                }
            }
        },
        "/planes-estudio": {
            "get": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Obtener todos los planes de estudio",
                "operationId": "5fda4ae9065ce1e21ffa897d5d5139ee",
                "responses": {
                    "200": {
                        "description": "Planes de estudio obtenidos con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_carrera": {
                                                "type": "integer"
                                            },
                                            "id_modalidad": {
                                                "type": "integer"
                                            },
                                            "materias": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id_materia": {
                                                            "type": "integer"
                                                        },
                                                        "id_cat_nivel": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Crear un nuevo plan de estudio",
                "operationId": "a7daba212b8ff74f82b053a30c16b91f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_carrera",
                                    "id_modalidad",
                                    "materias"
                                ],
                                "properties": {
                                    "id_carrera": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_modalidad": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "materias": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "id_materia",
                                                "id_cat_nivel"
                                            ],
                                            "properties": {
                                                "id_materia": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_cat_nivel": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plan de estudio creado con éxito"
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    },
                    "500": {
                        "description": "Error al crear el plan de estudio"
                    }
                }
            }
        },
        "/planes-estudio/{id_plan_estudio}": {
            "put": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Actualizar un plan de estudio existente",
                "operationId": "6f799e7ad0726dfcbccdbc1b07066019",
                "parameters": [
                    {
                        "name": "id_plan_estudio",
                        "in": "path",
                        "description": "ID del plan de estudio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_carrera": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_modalidad": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_materia": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cat_nivel": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plan de estudio actualizado con éxito"
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    },
                    "500": {
                        "description": "Error al actualizar el plan de estudio"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Eliminar un plan de estudio",
                "operationId": "959ec023c359062e6656cd0c3a6ac27b",
                "parameters": [
                    {
                        "name": "id_plan_estudio",
                        "in": "path",
                        "description": "ID del plan de estudio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plan de estudio eliminado con éxito"
                    },
                    "500": {
                        "description": "Error al eliminar el plan de estudio"
                    }
                }
            }
        },
        "/planes-estudio/materia": {
            "delete": {
                "tags": [
                    "Planes de Estudio"
                ],
                "summary": "Eliminar una materia de un plan de estudio",
                "operationId": "6d292ad9bf93ef94a2e2c69f2b881df9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_plan_estudio",
                                    "id_cat_nivel",
                                    "id_materia"
                                ],
                                "properties": {
                                    "id_plan_estudio": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_cat_nivel": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_materia": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materia eliminada del plan de estudio con éxito"
                    },
                    "422": {
                        "description": "Datos de entrada no válidos"
                    },
                    "500": {
                        "description": "Error al eliminar la materia del plan de estudio"
                    }
                }
            }
        },
...
    },
...
}